# factor out bin calculation and optimize push!()