--- a/src/hist1d.jl
+++ b/src/hist1d.jl
@@ -69,6 +69,9 @@ end
 Adding one value at a time into histogram. 
 `sumw2` (sum of weights^2) accumulates `wgt^2` with a default weight of 1.
 `unsafe_push!` is a faster version of `push!` that is not thread-safe.
+
+N.B. To append multiple values at once, use broadcasting via
+`push!.(h, [-3.0, -2.9, -2.8])` or `push!.(h, [-3.0, -2.9, -2.8], 2.0)`
 """
 @inline function Base.push!(h::Hist1D{T,E}, val::Real, wgt::Real=1.0) where {T,E}
     r = h.hist.edges[1]
@@ -99,6 +102,8 @@ end
     @inbounds h.sumw2[binidx] += c*wgt^2
     return nothing
 end
+
+Base.broadcastable(h::Hist1D) = Ref(h)
 
 """
     Hist1D(elT::Type{T}=Float64; binedges) where {T}