Closes https://github.com/Moelf/FHist.jl/issues/11

* Factored out bin calculation logic for float range, int range, irregular binning.
* Added `empty!()` to reset hists
* Added thread-UNsafe `unsafe_push!()`
* Defined `Base.broadcastable(h::Hist1D) = Ref(h)` so one can do `push!.(h, [-3.0, -2.9, -2.8])` or `push!.(h, [-3.0, -2.9, -2.8], 2.0)` as a "broadcasted push".

### Master
```julia
julia> using Revise, FHist, BenchmarkTools, StatsBase, StatProfilerHTML

julia> a = randn(10^6);

# one-shot construction
julia> @benchmark Hist1D(a, -3:0.1:3)
BechmarkTools.Trial: 1239 samples with 1 evaluations.
 Range (min … max):  3.577 ms …   5.555 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.957 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.024 ms ± 271.173 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 1.31 KiB, allocs estimate: 6.

julia> h = Hist1D(Int; bins=-3:0.1:3);

# thread-safe iterative construction
julia> @benchmark for i in $a push!($h, i) end
BechmarkTools.Trial: 70 samples with 1 evaluations.
 Range (min … max):  67.342 ms … 92.730 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     70.394 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   71.623 ms ±  4.372 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 0 bytes, allocs estimate: 0.
```
### This PR

```julia
# one-shot construction
julia> @benchmark Hist1D(a, 3:0.1:3)
 Range (min … max):  3.271 ms …   5.076 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.472 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.511 ms ± 160.078 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

# thread-safe iterative construction
julia> @benchmark for i in $a push!($h, i) end
 Range (min … max):  27.190 ms …  33.794 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     28.067 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   28.241 ms ± 920.786 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 0 bytes, allocs estimate: 0.

# thread-unsafe iterative construction
julia> @benchmark for i in $a unsafe_push!($h, i) end
 Range (min … max):  4.251 ms …   6.034 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.563 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.601 ms ± 189.220 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
```

So, a ~2.5x speedup wrt master for `push!()` and apparently `@inline` + removing locks gives another ~6x speedup! `unsafe_push!()` is roughly the same speed as the one-shot method.