@Moelf 2021-08-04
# [Codecov](https://codecov.io/gh/Moelf/FHist.jl/pull/12?src=pr&el=h1&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling) Report
> Merging [#12](https://codecov.io/gh/Moelf/FHist.jl/pull/12?src=pr&el=desc&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling) (d8b44c7) into [master](https://codecov.io/gh/Moelf/FHist.jl/commit/6a84ae93db20b1243bc25647e483d75fdd52d607?el=desc&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling) (6a84ae9) will **decrease** coverage by `5.09%`.
> The diff coverage is `55.55%`.

[![Impacted file tree graph](https://codecov.io/gh/Moelf/FHist.jl/pull/12/graphs/tree.svg?width=650&height=150&src=pr&token=TBQ6VFHB1Y&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling)](https://codecov.io/gh/Moelf/FHist.jl/pull/12?src=pr&el=tree&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling)

```diff
@@            Coverage Diff             @@
##           master      #12      +/-   ##
==========================================
- Coverage   67.51%   62.42%   -5.10%     
==========================================
  Files           5        5              
  Lines         157      157              
==========================================
- Hits          106       98       -8     
- Misses         51       59       +8     
```


| [Impacted Files](https://codecov.io/gh/Moelf/FHist.jl/pull/12?src=pr&el=tree&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling) | Coverage Δ | |
|---|---|---|
| [src/hist1d.jl](https://codecov.io/gh/Moelf/FHist.jl/pull/12/diff?src=pr&el=tree&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling#diff-c3JjL2hpc3QxZC5qbA==) | `75.51% <55.55%> (-8.17%)` | :arrow_down: |

------

[Continue to review full report at Codecov](https://codecov.io/gh/Moelf/FHist.jl/pull/12?src=pr&el=continue&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling).
> **Legend** - [Click here to learn more](https://docs.codecov.io/docs/codecov-delta?utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling)
> `Δ = absolute <relative> (impact)`, `ø = not affected`, `? = missing data`
> Powered by [Codecov](https://codecov.io/gh/Moelf/FHist.jl/pull/12?src=pr&el=footer&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling). Last update [6a84ae9...d8b44c7](https://codecov.io/gh/Moelf/FHist.jl/pull/12?src=pr&el=lastupdated&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling). Read the [comment docs](https://docs.codecov.io/docs/pull-request-comments?utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=Jerry+Ling).